added validations for update and delete

+++ src/order/order.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, InternalServerErrorException, Logger, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, InternalServerErrorException, Logger, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateOrderDto } from './dto/create-order.dto';
 import {  UpdateOrderItemDto } from './dto/update-order.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -6,6 +6,8 @@ import { Order } from './entities/order.entity';
 import { Repository } from 'typeorm';
 import { OrderItem } from './entities/orderItem.entity';
 import { CreateOrderItemDto } from './dto/create-orderItem.dto';
+import { User, UserRole } from 'src/user/entities/user.entity';
+import { use } from 'passport';
 
 @Injectable()
 export class OrderService {
@@ -41,22 +43,29 @@ export class OrderService {
     return await this.orderItemRepository.findOneBy({id:id});
   }
 
-  async update(id: string, updateOrderItemDto: UpdateOrderItemDto) {
-    const order = await this.orderItemRepository.preload({
-      id: id,
-      ...updateOrderItemDto
-    });
+  async update(req:any,id: string, updateOrderItemDto: UpdateOrderItemDto) {
+    const user = req.user as User;
+    const orderItem :OrderItem = await this.orderItemRepository.findOneBy({id:id})
+    const order : Order = await this.ordersRepository.findOneBy({id:orderItem.orderId})
+    if (order.buyerId==user.id||user.roles.includes(UserRole.ADMIN)){ 
+      const order = await this.orderItemRepository.preload({
+        id: id,
+        ...updateOrderItemDto
+      });
 
-    if ( !order ) throw new NotFoundException(`Order Item with id: ${ id } not found`);
+      if ( !order ) throw new NotFoundException(`Order Item with id: ${ id } not found`);
 
-    try {
-      await this.orderItemRepository.save( order );
-      return order;
-      
-    } catch (error) {
-      this.handleDBExceptions(error);
+      try {
+        await this.orderItemRepository.save( order );
+        return order;
+        
+      } catch (error) {
+        this.handleDBExceptions(error);
+      }
+      return `This action updates a #${id} order`;
+    }else{
+      throw new UnauthorizedException;
     }
-    return `This action updates a #${id} order`;
   }
   private handleDBExceptions( error: any ) {
 
@@ -69,12 +78,27 @@ export class OrderService {
 
   }
 
-  async removeOrder(id: string) {
-    const order = await this.findOne(id);
-    return await this.ordersRepository.remove(order);
+  async removeOrder(req:any, id: string) { 
+    const user = req.user as User;
+    const orderItem :OrderItem = await this.orderItemRepository.findOneBy({id:id})
+    const order : Order = await this.ordersRepository.findOneBy({id:orderItem.orderId})
+    if (order.buyerId==user.id||user.roles.includes(UserRole.ADMIN)){    
+      const order = await this.findOne(id);
+      return await this.ordersRepository.remove(order);
+    }
+    else{
+      throw new UnauthorizedException;
+    }
   }
-  async removeOrderItem(id: string) {
+  async removeOrderItem(req:any, id: string) {
+    const user = req.user as User;
+    const orderItem :OrderItem = await this.orderItemRepository.findOneBy({id:id})
+    const order : Order = await this.ordersRepository.findOneBy({id:orderItem.orderId})
+    if (order.buyerId==user.id||user.roles.includes(UserRole.ADMIN)){ 
     const item = await this.findOneItem(id);
     return await this.orderItemRepository.remove(item);
+    }else{
+      throw new UnauthorizedException;
+    }
   }
 }
